pyproject.toml
[tool.poetry]
name = "codechain"
version = "0.1.0"
description = ""
authors = ["Fazio Alessandro 3 (EXT Spindox) <alessandro.fazio@spindox.it>"]
readme = "README.md"
packages = [{ include = "codechain", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
cloup = "^3.0.8"
numpy = "==2.2.0"
numba = "^0.61.2"
pydantic = "^2.11.7"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
codechain = "codechain.cli.codechain:run"
